refactor(agents): migrate internal imports to public API paths
Replace internal module path imports with package-level public imports
in the agents module. This improves API consistency and reduces coupling
to internal implementation details.

Changes:
- agents/__init__.py: models.prompt_meta → models
- agents/base.py: split models.review_runner import (AgentOptions public, AgentResult internal)
- agents/models.py: config.settings → config
- agents/plan_prompt.py: 6 imports migrated to public paths
- agents/review_pipeline_helpers.py: analysis.snapshot_diff → analysis, models.snapshot → models
- agents/review_prompt.py: 6 imports migrated to public paths
- agents/run_prompt.py: 5 imports migrated to public paths
- agents/backends/__init__.py: config.agent_preset → config
- agents/backends/codeagent.py: split config.agent_preset and models.review_runner imports
- agents/backends/codeagent_config.py: split config.agent_preset import

Exempted imports (9 types, upstream not exported):
- AgentResult, PlanRequest, ReviewRequest, AsyncExecutionHandle
- snapshot_service symbols, build_snapshot_diff_section
- codeagent_helpers symbols
- config.agent_preset utility functions

Note: Pre-commit mypy errors are pre-existing (not introduced by this change).
