+++ src/vibe3/agents/__init__.py
@@ -72,7 +72,7 @@ from vibe3.agents.run_prompt import (
     make_run_context_builder,
     make_skill_context_builder,
 )
-from vibe3.models.prompt_meta import PromptContextMode
+from vibe3.models import PromptContextMode
 
 __all__ = [
     # Protocols & Models
