+++ src/vibe3/agents/backends/codeagent_config.py
@@ -9,12 +9,9 @@ from typing import Any, Final
 
 from loguru import logger
 
-from vibe3.config.agent_preset import (
-    read_models_json,
-    repo_models_json_path,
-    resolve_effective_agent_options,
-)
-from vibe3.models.review_runner import AgentOptions
+from vibe3.config import resolve_effective_agent_options
+from vibe3.config.agent_preset import read_models_json, repo_models_json_path
+from vibe3.models import AgentOptions
 
 # Path to codeagent models config
 MODELS_JSON_PATH: Final[Path] = Path.home() / ".codeagent" / "models.json"
