+++ src/vibe3/agents/base.py
@@ -1,7 +1,8 @@
 from pathlib import Path
 from typing import Protocol, runtime_checkable
 
-from vibe3.models.review_runner import AgentOptions, AgentResult
+from vibe3.models import AgentOptions
+from vibe3.models.review_runner import AgentResult
 
 
 @runtime_checkable
