+++ src/vibe3/agents/review_pipeline_helpers.py
@@ -7,13 +7,13 @@ from typing import Any
 import typer
 from loguru import logger
 
-from vibe3.analysis.snapshot_diff import compute_diff
+from vibe3.analysis import compute_diff
 from vibe3.analysis.snapshot_service import (
     SnapshotError,
     build_snapshot,
     find_snapshot_by_branch,
 )
-from vibe3.models.snapshot import StructureDiff
+from vibe3.models import StructureDiff
 
 
 def run_inspect_json(args: list[str]) -> dict[str, object]:
